Refactored Main activity
Refactored status out of main activity. Add MVC implementation, moved all the status methods into the feed service. Implemented status observer.
