+++ app/src/main/java/edu/byu/cs/tweeter/client/view/main/MainActivity.java
@@ -4,7 +4,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -21,25 +20,16 @@ import androidx.viewpager.widget.ViewPager;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.tabs.TabLayout;
 
-import java.net.MalformedURLException;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 import edu.byu.cs.client.R;
 import edu.byu.cs.tweeter.client.backgroundTask.IsFollowerTask;
-import edu.byu.cs.tweeter.client.backgroundTask.PostStatusTask;
 import edu.byu.cs.tweeter.client.cache.Cache;
 import edu.byu.cs.tweeter.client.presenter.MainPresenter;
 import edu.byu.cs.tweeter.client.view.login.LoginActivity;
 import edu.byu.cs.tweeter.client.view.login.StatusDialogFragment;
 import edu.byu.cs.tweeter.client.view.util.ImageUtils;
-import edu.byu.cs.tweeter.model.domain.Status;
 import edu.byu.cs.tweeter.model.domain.User;
 
 /**
@@ -157,81 +147,7 @@ public class MainActivity extends AppCompatActivity implements StatusDialogFragm
   public void onStatusPosted(String post) {
     postingToast = Toast.makeText(this, "Posting Status...", Toast.LENGTH_LONG);
     postingToast.show();
-
-    try {
-      Status newStatus = new Status(post, Cache.getInstance().getCurrUser(), getFormattedDateTime(), parseURLs(post), parseMentions(post));
-      PostStatusTask statusTask = new PostStatusTask(Cache.getInstance().getCurrUserAuthToken(),
-              newStatus, new PostStatusHandler());
-      ExecutorService executor = Executors.newSingleThreadExecutor();
-      executor.execute(statusTask);
-    } catch (Exception ex) {
-      Log.e(LOG_TAG, ex.getMessage(), ex);
-      Toast.makeText(this, "Failed to post the status because of exception: " + ex.getMessage(), Toast.LENGTH_LONG).show();
-    }
-  }
-
-  public String getFormattedDateTime() throws ParseException {
-    SimpleDateFormat userFormat = new SimpleDateFormat("yyyy-MM-dd hh:mm:ss");
-    SimpleDateFormat statusFormat = new SimpleDateFormat("MMM d yyyy h:mm aaa");
-
-    return statusFormat.format(userFormat.parse(LocalDate.now().toString() + " " + LocalTime.now().toString().substring(0, 8)));
-  }
-
-  public List<String> parseURLs(String post) throws MalformedURLException {
-    List<String> containedUrls = new ArrayList<>();
-    for (String word : post.split("\\s")) {
-      if (word.startsWith("http://") || word.startsWith("https://")) {
-
-        int index = findUrlEndIndex(word);
-
-        word = word.substring(0, index);
-
-        containedUrls.add(word);
-      }
-    }
-
-    return containedUrls;
-  }
-
-  public List<String> parseMentions(String post) {
-    List<String> containedMentions = new ArrayList<>();
-
-    for (String word : post.split("\\s")) {
-      if (word.startsWith("@")) {
-        word = word.replaceAll("[^a-zA-Z0-9]", "");
-        word = "@".concat(word);
-
-        containedMentions.add(word);
-      }
-    }
-
-    return containedMentions;
-  }
-
-  public int findUrlEndIndex(String word) {
-    if (word.contains(".com")) {
-      int index = word.indexOf(".com");
-      index += 4;
-      return index;
-    } else if (word.contains(".org")) {
-      int index = word.indexOf(".org");
-      index += 4;
-      return index;
-    } else if (word.contains(".edu")) {
-      int index = word.indexOf(".edu");
-      index += 4;
-      return index;
-    } else if (word.contains(".net")) {
-      int index = word.indexOf(".net");
-      index += 4;
-      return index;
-    } else if (word.contains(".mil")) {
-      int index = word.indexOf(".mil");
-      index += 4;
-      return index;
-    } else {
-      return word.length();
-    }
+    presenter.postStatus(post, Cache.getInstance().getCurrUser());
   }
 
   public void updateFollowButton(boolean removed) {
@@ -273,26 +189,6 @@ public class MainActivity extends AppCompatActivity implements StatusDialogFragm
       }
     }
   }
-
-  // PostStatusHandler
-
-  private class PostStatusHandler extends Handler {
-    @Override
-    public void handleMessage(@NonNull Message msg) {
-      boolean success = msg.getData().getBoolean(PostStatusTask.SUCCESS_KEY);
-      if (success) {
-        postingToast.cancel();
-        Toast.makeText(MainActivity.this, "Successfully Posted!", Toast.LENGTH_LONG).show();
-      } else if (msg.getData().containsKey(PostStatusTask.MESSAGE_KEY)) {
-        String message = msg.getData().getString(PostStatusTask.MESSAGE_KEY);
-        Toast.makeText(MainActivity.this, "Failed to post status: " + message, Toast.LENGTH_LONG).show();
-      } else if (msg.getData().containsKey(PostStatusTask.EXCEPTION_KEY)) {
-        Exception ex = (Exception) msg.getData().getSerializable(PostStatusTask.EXCEPTION_KEY);
-        Toast.makeText(MainActivity.this, "Failed to post status because of exception: " + ex.getMessage(), Toast.LENGTH_LONG).show();
-      }
-    }
-  }
-
 
   @Override
   public void followButtonUpdates(boolean value) {
